Improved diagnostics
As you type, you emit a lot of invalid code, and I was not catching
the errors that Code.compile_string was throwing, which caused the
remote control app to die and cause completion and compilation to go
away.
In addition, the error handling of message strings was not resilient
enough to handle the variety of error types that elixir can
create. The errors module now has improved error parsing that should
be a lot more tolerant.