Fixed pyramid module suggestions
Code completion for modules uses snippets, which insert syntactically valid code
into the current editor. In combination with compile-as-you-type, this
caused modules to be generated for each character you typed in the
module name. This meant that if you named your module "Dog", lexical
would create "D", "Do" and "Dog" modules in your project.

This PR uses an idea of an "empty" module, which is a module that
defines no functions, no macros, no structs and no types. If the
module that's being compiled is empty in this way, then it's purged
and deleted from the project server.

Unfortunately, this meant we had to move off of Code.eval_quoted, as
that function only returns the last thing being updated, which meant
that a file that created several modules would only have the top-level
module returned. Code.compile_quoted, on the other hand, returns all
the modules that have been compiled in the quoted AST. Unfortunately,
Code.compile_quoted doesn't return the line numbers in a number of
error cases, so we had to go to great lengths to resurrect them.