Improved struct completion
Struct completion now takes the surrounding context into account, and
provides completions based upon it.
If the context starts with '%', completions will only return struct
modules or parents of struct modules, then when a module is selected,
if you have a nested module like %Foo.Bar.Baz, the completion doesn't include the
leading '%', but does include the trailing '{}'. If, however, you have
a single module, both the leading '%' and trailing '{}' are included.

However, if the context doesn't start with a '%', only the module name
is used in the completion.