Reworked protocol encoding/decoding
I was noticing that we had a lot of code that was doing line and
character conversion inline. One example of this is the encoding work in
the diagnostics code. In order to fix this, I removed the ranged
protocol in favor of two new protocols, DocumentContainer and
Convertible.

A document container holds things that all refer to the same document.
This is necessary because having the document is necesary for some
conversions (ranges, positions) to occur. Once a document has been
determined, conversions are up to the Convertible protocol. Prior to
being written, structures (and lists and maps) are subjected to the
Convertible protocol to ensure that everything emitted is valid LSP.

The effect of this is to relegate most of the LSP structures to a very
small part of the code, as implementers will only need to deal with
Lexical structures for the most part. One of the benefits though is that
one can still embed lexical structures inside of LSP structures, and
they will be converted automatically, as long as they're inside some
kind of document container.

To this end, I added the following source file structures:

`Edit`: represents a singlen edit to a document. An analog of TextEdit
`DocumentEdits`: Represents a set of changes to a document.

Code mods should now speak in edits and document edits rather than
TextEdits.

Expect more changes in the future to futher move in this direction.