Plugin architecture
This is the first shot at the plugin architecture that I think will work
across projects. The goals of the architecture are:

  1. Plugins should never be able to crash lexical
  2. Plugins should be able to be embedded in other projects, and be
     discovered by lexical
  3. Plugins can be created for different aspects of lexical with a minimal set of dependencies

I think I've done it, to the point where it's possible to include a
simple plugin for diagnostics in the plugin documentation.

Some design considerations that I feel could use some thought:

 * While it's not possible for a plugin to crash lexial, a constantly
   crashing plugin would be annoying, so I disabled plugins after 10
   crashes over any time frame. This might be too aggressive.
 * Plugins will be separated by the kind of plugin they are, so it will
   be difficult, for example, to have a plugin both be a Diagnostic and
   a Code Action plugin.

There are a lot of changes here, but this required moving some code
around. Feel free to ignore the moves.