Namespaced proto and protocol apps (#268)
The protocol and proto apps were not being namespaced because I missed
appending a trailing '.' to the root module names, this meant both
were being excluded.

Circling back, I don't know why I was excluding any of the deps apps
from namespacing, as it seems to work fine with it.

It was also possisble to apply namespacing to modules that were already
namespaced, so I added a check to ensure that doesn't happen. This was
happening for plugins in external projects, as the plugin is a dep, but
the code is written to the .lexical directory.