Fixed bug in path namespacing (#350)
I was using String.replace to apply namespacing to a path. This would
cause issues if lexical resided in a directory whose parents had a
substring of one of its app names (e.g. language_server).

Instead of using replace, we should only apply namespacing to
subdirectories that are exactly equal to one of lexical's apps.

Fixes #341