Always return `Completion.List` with `is_incomplete: true`
The `:is_incomplete` flag signals to the client that completions should
be requested on every keypress. If it is not set, the client will
generally issue a single completion request when the user starts typing
a new "word", and then will filter client-side as subsequent letters are
typed.

This causes problems for things like `def foo do|` -- VS Code, for
instance, issues the request after the `d` of `do` is typed, and does
not make another when the `o` is typed. This means you do not see the
special-cased completion for a `do/end` block, but instead see a
filtered list of completions for everything starting with `d`.