Reorder the startup order of the children of `Server.Project.Supervisor` (#407)
After we moved the dispatch, the first progress message appeared very slowly, taking about **10** seconds. 
This PR primarily aims to start the `progress` as the first process in `Lexical.Server.Project.Supervisor`, which will advance the appearance of the first progress message by approximately *4* seconds.

In addition, we only need to register the *progress listener* when we start both nodes, so I pass `progress: true` to dispatch's `start_link` in the `application.ex` of the `remote_control`.