Refactor completions to always emit text edits (#409)
There are two basic types of completion proposal that language servers can emit: `insertText` and `textEdit`. The primary difference is that `textEdit` specifies the exact range to be replaced by the completion, while it's up to the editor to determine the range for `insertText`. This is usually done on "word boundary," which can differ from editor to editor based on the grammar they're using.

Using `insertText` is especially problematic for completions with prefixes, like `%`, `:`, `@`, etc. If you were completing `@com|` and accepted `@completion` using `insertText`, some editors would only replace the prefix `com`, resulting in `@@completion`.

Always emitting `textEdit` with an explicit range allows us to solve for those issues.

Closes #351