Fix crash on boot due to precedence bug and stdio log (#444)
The checks in `activate_version_manager.sh` need explicit `{ }` to prevent `_detect_rtx` from always running. This would cause the script to exit with an error spuriously, which would trigger an error message that was being written to stdio. Because we use stdio as our transport, that would cause the LS client to crash.

This fixes the crash reported in #440. However, it will still run the server using asdf if both asdf and rtx are installed with the correct versions of Elixir.