Added reindex command (#522)
* Added reindex command

Added a reindex code lens so we can rebuild indexes without resorting
to manual operations.

As we build the indexing infrastructure out, we'll likely need to
rebuild the index a lot. Presently, this means restarting the server,
which can take some time, so I thought i'd add a command to do it
instead. I tried using code actions, but this was a bit fraught, so
instead, I used a code lens on the project definition in your mix.exs file.
The code lens disappears when clicked and reappears when the indexing job is done.