ETS backend write-ahead log (#554)
* WAL backed store

Our store was using :ets.tab2file to write its entries to the
filesystem, which would require writing the entire index every time it
was called. This is problematic, as updates to the index come in
whenever code is added or removed. This would also cause a lot of
contention during startup because of the amount of data being read off
the filesystem.

This branch introduces a write ahead log, using erlang's built in
`disk_logger` module. The idea is that the WAL writes operations to
disk before they're applied to ETS, so that when we start up, we first
read a checkpoint, then apply the operations one-by-one to get to the
prior state.

Note that this approach is in contrast to using Khepri. Khepri is very
nice, but it's significantly slower than this approach, and it uses a
lot more memory. It's also worth noting that the WAL based approach
can have its memory efficiency improved by removing the need to
shuffle all of the entries around between processes, but that's future
work.