Simplifications in `remote_control` (#778)
* Remove dead code in `RemoteControl.Application`

* More conventional init arg for `RemoteControl.Commands.Reindex`

* Always enable config compiler

We don't support Elixir < 1.11, so the feature check is unnecessary.

* Always use deps.safe_compile, which calls deps.compile if possible

* Remove dead code in `ModuleMappings` and test correct code path

`ModuleMappings.update/2` was only used in the tests. I've refactored them to use dispatched messages, which is how the mappings are actually updated during regular use.

* Delete `RemoteControl.Modules.Predicate.Syntax`

It was only used for two calls and was difficult to understand. Instead, we now explicitly pass an mfa tuple to `RemoteControl.Modules.with_prefix/2` and the module being testing is added as the first arg.

* Delete unused `RemoteControl.ProjectNode.Launcher`