Don't run top-level code in `.exs` files (#798)
Lexical's handling of `.exs` files left something to be desired (understatement). If you had a top-level Benchee benchmark, for instance, Lexical would run it as part of it's as-you-type compilation. Oops.

This change dramatically improves Lexical's `.exs` handling by wrapping unsafe top-level forms in a generated module and def, which still allows compile-time warnings and errors to come through without actually running that top-level code during compilation.

There are two known issues with this approach:

1. The following will report an error in-editor, even though it runs fine separately:

```elixir
my_module = My.Module
defmodule my_module do
  ...
end
```

As this is an extremely uncommon pattern, we're accepting this as a known defect and have no plans to fix at the moment.

2. `Mix.install` remains unsupported and completions/warnings/errors/etc. will appear as if the file is running in the context of the containing Mix project. Support for `Mix.install` is tracked in #403.