fix(engine): support go to definition when function is called via __MODULE__ (#261)
This wasn't working, as __MODULE__ is recognized as :var, not as
:module in AST. What was needed is to add a special handling of it in
`expand_alias` function.

Recognizing __MODULE__ as :var, while semantically incorrect, does not
seem to be a problem in scope of definition finding. It's not strictly a
module either, so to be super-correct, we would need to add a special
handling just for it.